# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: github.com/is7qin/c3api Admin API
  version: 1.0.0
  description: '管理端 API（模板/账号/分组/规则/日志/统计）。认证：Authorization: Bearer <admin_token>。'
servers:
  - url: /admin
...
paths:
...
  /templates/{id}/ext:
    parameters:
      - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
    get:
      operationId: GetTemplatesIdExt
      tags: [templates]
      summary: 读取模板类型化扩展（编辑回显；仅生态三类型模板有 ext 行）
      responses:
        '200':
          description: 模板 ext 配置
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TemplateExt' }
        default:
          $ref: '#/components/responses/Error'
    put:
      operationId: PutTemplatesIdExt
      tags: [templates]
      summary: 幂等写入模板类型化扩展（Create/Update 合一；全列更新含 NULL 清空）
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TemplateExt' }
      responses:
        '200':
          description: 写入后的 ext 配置
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TemplateExt' }
        default:
          $ref: '#/components/responses/Error'
  /accounts:
    post:
      operationId: PostAccounts
      tags: [accounts]
      summary: 创建账号
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AccountCreate' }
      responses:
        '200':
          description: 创建后的账号
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Account' }
        default:
          $ref: '#/components/responses/Error'
    get:
      operationId: GetAccounts
      tags: [accounts]
      summary: 账号列表（分页/筛选/排序，含运行时视图）
      parameters:
        - { name: limit, in: query, schema: { type: integer, default: 20, maximum: 200, description: 每页条数（上限 200，超限裁剪到 200） } }
        - { name: offset, in: query, schema: { type: integer, default: 0 } }
        - { name: name, in: query, schema: { type: string } }
        - { name: sort, in: query, schema: { type: string } }
        - { name: order, in: query, schema: { type: string, enum: [asc, desc] } }
        - { name: status, in: query, schema: { type: string, description: "多值逗号分隔：active,unhealthy,429,disabled" } }
        - { name: template_id, in: query, schema: { type: integer, format: int64 } }
      responses:
        '200':
          description: 账号视图列表
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AccountListResponse' }
        default:
          $ref: '#/components/responses/Error'
  /accounts/batch-delete:
    post:
      operationId: PostAccountsBatchDelete
      tags: [accounts]
      summary: 批量删除账号（事务，全成或全败）
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BatchDeleteBody' }
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BatchDeleteResponse' }
        default:
          $ref: '#/components/responses/Error'
  /accounts/batch-update:
    post:
      operationId: PostAccountsBatchUpdate
      tags: [accounts]
      summary: 批量更新账号（fields 为任意字段子集）
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BatchUpdateAccountsBody' }
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BatchUpdateResponse' }
        default:
          $ref: '#/components/responses/Error'
  /accounts/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
    get:
      operationId: GetAccountsId
      tags: [accounts]
      responses:
        '200':
          description: 账号
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Account' }
        default:
          $ref: '#/components/responses/Error'
    put:
      operationId: PutAccountsId
      tags: [accounts]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AccountCreate' }
      responses:
        '200':
          description: 更新后账号
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Account' }
        default:
          $ref: '#/components/responses/Error'
    delete:
      operationId: DeleteAccountsId
      tags: [accounts]
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DeletedResponse' }
        default:
          $ref: '#/components/responses/Error'
  /accounts/{id}/groups:
    parameters:
      - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
    get:
      operationId: GetAccountsIdGroups
      tags: [accounts]
      summary: 读取账号的全部分组 id（编辑回显；不随账号列表返回）
      responses:
        '200':
          description: 分组 id 列表
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AccountGroupsResponse' }
        default:
          $ref: '#/components/responses/Error'
  /accounts/{id}/ext:
    parameters:
      - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
    get:
      operationId: GetAccountsIdExt
      tags: [accounts]
      summary: 读取账号类型化鉴权扩展（编辑回显；仅 codex-oauth/codex-pat 账号有 ext 行）
      responses:
        '200':
          description: 账号 ext 配置
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AccountExt' }
        default:
          $ref: '#/components/responses/Error'
    put:
      operationId: PutAccountsIdExt
      tags: [accounts]
      summary: 幂等写入账号类型化鉴权扩展（Create/Update 合一；全列更新含 NULL 清空）
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AccountExt' }
      responses:
        '200':
          description: 写入后的 ext 配置
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AccountExt' }
        default:
          $ref: '#/components/responses/Error'
...
components:
  responses:
    Error:
      description: 错误响应
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: { type: string }
    DeletedResponse:
      type: object
      required: [deleted]
      properties:
        deleted: { type: boolean }
...
    AccountStatus:
      type: string
      enum: [active, unhealthy, "429", disabled]
...
    Template:
      type: object
      required: [ID, Name, BaseURL, SupportedFormats, CreatedAt, UpdatedAt]
      properties:
        ID: { type: integer, format: int64 }
        Name: { type: string }
        BaseURL: { type: string }
        CredentialType: { type: string, enum: [api_key, responses-special, codex-oauth, codex-pat], description: 模板号池类型；生态三类型只支持 resp / resp-ws 格式 }
        SupportedFormats:
          type: array
          items: { type: string, enum: [openai-chat, openai-responses, openai-responses-ws, openai-images, openai-search, anthropic] }
        Models:
          type: array
          items: { type: string }
        FormatModels:
          type: object
          additionalProperties:
            type: array
            items: { type: string }
        ModelMapping:
          type: object
          additionalProperties: { type: string }
        CreatedAt: { type: string, format: date-time }
        UpdatedAt: { type: string, format: date-time }
        DeletedAt: { type: string, format: date-time, nullable: true, description: 软删除时间戳；null = 存活（列表过滤已删；GET 单个可查已删项） }
    AccountCreate:
      type: object
      required: [name, template_id, upstream_key]
      properties:
        name: { type: string }
        template_id: { type: integer, format: int64 }
        # 账号级 base_url 覆盖（路由属性，位于 upstream_key 前——用户裁决
        # 2026-08-14）：非空 = 覆盖模板 base_url；""/null/缺省统一 = 继承模板
        # （create 路径空串归一 nil，与批量三态语义分写）。
        base_url: { type: string, nullable: true, description: "账号级覆盖（可选）：非空优先于模板 base_url；null = 继承模板" }
        # upstream_key：api_key/responses-special 模板必填（静态透传上游 key）；
        # codex-oauth/codex-pat 模板可选（凭据走 account_ext——创建后经
        # /accounts/{id}/ext 配置；service 按模板类型校验）。
        upstream_key: { type: string }
        status: { $ref: '#/components/schemas/AccountStatus' }
        weight: { type: integer }
        max_concurrency: { type: integer }
        # 创建时可选：null/缺省 = 无分组；提供 = 替换语义（空数组 = 无分组）
        group_ids:
          type: array
          items: { type: integer, format: int64 }
    Account:
      type: object
      properties:
        ID: { type: integer, format: int64 }
        Name: { type: string }
        TemplateID: { type: integer, format: int64 }
        Template: { $ref: '#/components/schemas/Template' }
        # 账号级 base_url 覆盖（路由属性，位于 upstream_key 前）：null = 继承模板
        BaseURL: { type: string, nullable: true, description: "账号级覆盖：非空优先于模板 base_url；null = 继承模板" }
        UpstreamKey: { type: string }
        Status: { $ref: '#/components/schemas/AccountStatus' }
        CooldownUntil: { type: string, format: date-time, nullable: true }
        Weight: { type: integer }
        MaxConcurrency: { type: integer }
        LastError: { type: string, nullable: true }
        LastUsedAt: { type: string, format: date-time, nullable: true }
        CreatedAt: { type: string, format: date-time }
        UpdatedAt: { type: string, format: date-time }
        DeletedAt: { type: string, format: date-time, nullable: true, description: 软删除时间戳；null = 存活（列表过滤已删；GET 单个可查已删项） }
    AccountView:
      allOf:
        - $ref: '#/components/schemas/Account'
        - type: object
          properties:
            concurrency: { type: integer, format: int64 }
            err_rate: { type: number, format: double }
            err_count: { type: integer }
    TemplateExt:
      type: object
      required: [credential_type]
      properties:
        # template_id 只读（响应）；请求体忽略——路径 {id} 为准。
        template_id: { type: integer, format: int64 }
        credential_type: { type: string, enum: [responses-special, codex-oauth, codex-pat], description: "类型声明；必须与父模板 credential_type 一致（不一致 → 400）" }
        # 模板是共享配置面：唯一可配置列 strip_image_tools（三类型公共能力开关）——
        # 凭据列组（oauth/pat）一律在账号级 account_ext（账号私有）。
        strip_image_tools: { type: boolean, nullable: true, description: "三类型公共能力开关：模板级图像 tool 剥离（NULL = 未配置 = 关闭）" }
    AccountExt:
      type: object
      required: [credential_type]
      properties:
        # account_id 只读（响应）；请求体忽略——路径 {id} 为准。
        account_id: { type: integer, format: int64 }
        credential_type: { type: string, enum: [codex-oauth, codex-pat], description: 类型-列组约束（service 校验）：oauth 只允许 oauth_* 列组；pat 只允许 pat_key }
        # 身份四元组（真实 codex 客户端语义；首次写入缺省 → service 自动生成
        # （UUIDv4/v7 形状）并持久化，账号存在期间稳定；显式提供 → 采用；
        # 恒等 thread_id == session_id、window_id = {thread_id}:0）。
        installation_id: { type: string, description: "身份：账号级唯一（UUIDv4；响应恒有——首次写入自动生成）" }
        session_id: { type: string, nullable: true, description: "身份：会话级（UUIDv7；恒等 == thread_id；service 自动生成/沿用）" }
        thread_id: { type: string, nullable: true, description: "身份：会话级（UUIDv7；恒等 == session_id）" }
        window_id: { type: string, nullable: true, description: "身份：会话级派生 {thread_id}:0（导入时生成后恒定不变——恒 0，无透传无解析；上游不校验 n）" }
        # 凭据列组（按 credential_type 约束）
        oauth_token: { type: string, nullable: true, description: 凭据：oauth 访问令牌（oauth 行必填——最小完整性） }
        oauth_refresh_token: { type: string, nullable: true, description: 凭据：oauth 刷新令牌 }
        oauth_expires_at: { type: string, format: date-time, nullable: true, description: 凭据：oauth 访问令牌过期时间 }
        pat_key: { type: string, nullable: true, description: 凭据：pat }
        # 管理标识（非请求面）
        email: { type: string, nullable: true, description: "管理标识：codex 账号登录邮箱（导入时由人工/上游提供，非自动生成——NewCodexIdentity 只生成身份四元组；可空）" }
...
    AccountListResponse:
      type: object
      required: [total, rows]
      properties:
        total: { type: integer, format: int64 }
        rows:
          type: array
          items: { $ref: '#/components/schemas/AccountView' }
...
    AccountGroupsResponse:
      type: object
      required: [group_ids]
      properties:
        group_ids:
          type: array
          items: { type: integer, format: int64 }
    BatchDeleteBody:
      type: object
      required: [ids]
      properties:
        ids:
          type: array
          items: { type: integer, format: int64 }
          minItems: 1
          maxItems: 100
...
    AccountPatch:
      type: object
      properties:
        name: { type: string }
        template_id: { type: integer, format: int64 }
        # 批量三态定死（评审 C1）："" = 清空（落 NULL = 继承模板）；null/缺省 = 不变
        base_url: { type: string, nullable: true, description: "批量三态：空串 = 清空（回继承模板）；null/缺省 = 不变；非空 = 落值" }
        upstream_key: { type: string }
        status: { type: string, enum: [active, unhealthy, "429", disabled] }
        weight: { type: integer }
        max_concurrency: { type: integer }
        # 替换语义：null/缺省 = 不变；[] = 清空；提供列表 = 替换账号全部分组
        group_ids:
          type: array
          items: { type: integer, format: int64 }
...
    BatchUpdateAccountsBody:
      type: object
      required: [ids, fields]
      properties:
        ids:
          type: array
          items: { type: integer, format: int64 }
          minItems: 1
          maxItems: 100
        fields: { $ref: '#/components/schemas/AccountPatch' }
...
    BatchDeleteResponse:
      type: object
      required: [deleted]
      properties:
        deleted: { type: integer }
    BatchUpdateResponse:
      type: object
      required: [updated]
      properties:
        updated: { type: integer }
    # --- 兑换码（Phase 5 计费前基础设施：资源发放的通用载体） ---
...